+++ logprep/processor/calculator/processor.py
@@ -24,10 +24,9 @@ from functools import cached_property
 from pyparsing import ParseException
 
 from logprep.abc.processor import Processor
-from logprep.processor.base.exceptions import DuplicationError
 from logprep.processor.calculator.fourFn import BNF
 from logprep.processor.calculator.rule import CalculatorRule
-from logprep.util.helper import add_field_to, get_source_fields_dict
+from logprep.util.helper import get_source_fields_dict
 from logprep.util.decorators import timeout
 
 
@@ -83,14 +82,3 @@ class Calculator(Processor):
             return result
 
         return calculate(event, rule, expression)
-
-    def _write_target_field(self, event, rule, result):
-        add_successful = add_field_to(
-            event,
-            output_field=rule.target_field,
-            content=result,
-            extends_lists=rule.extend_target_list,
-            overwrite_output_field=rule.overwrite_target,
-        )
-        if not add_successful:
-            raise DuplicationError(self.name, [rule.target_field])